BFP_WEB/requirements.txt
pywebview>=4.4
reportlab>=3.6
openpyxl>=3.0
Pillow>=9.0
pyinstaller>=5.0
